refactor(P2.7 D1): dedup the WITH-key triplicate into one canonical with_clause_key
The three near-duplicate WITH-key helpers that P2.6 co-located in
render_plan/with_to_cte/mod.rs collapse to one canonical with_clause_key()
in src/utils/with_clause_key.rs (next to cte_naming, per REFACTORING_SAFETY_PLAN
§5.2 D1).

The D1 "verbatim triplicate" premise was half-wrong: get_with_key and
get_with_clause_key were byte-identical simple copies (exported_aliases →
sorted-join, else "with_var"), but generate_with_key_from_with_clause carried an
extra item-extraction fallback and was the version the authoritative
find_all_with_clauses_grouped relied on. Unified on that rich variant so grouping
and barrier-matching can never disagree on a WITH clause's identity.

Behavior-preserving, corpus-verified: corpus_sweep + sql_golden (529 tests)
byte-identical, proving the two simple copies never reached their divergent
branch on any corpus query. Also removed the now-orphaned nested
extract_with_alias (its logic moved into the util) and its two dead imports.

Gates: 1612 lib + 529 integration + 1 ratchet, 0 failures; goldens/corpus
byte-identical; ratchet net-zero; fmt/clippy clean; warning-free --tests build.
with_to_cte/mod.rs sheds 148 lines.
