+++ src/utils/mod.rs
@@ -3,3 +3,4 @@ pub mod cte_naming;
 pub mod id_encoding;
 pub mod serde_arc;
 pub mod serde_arc_vec;
+pub mod with_clause_key;
